refactor package location for pdf file
