+++ app/src/main/java/com/ferdinand/pdftestapp/data/models/PdfFile.kt
@@ -1,4 +1,4 @@
-package com.ferdinand.pdftestapp.data
+package com.ferdinand.pdftestapp.data.models
 
 import android.net.Uri
